add DnD to Pages and Folders
